server/requirements.txt
Django>=6.0,<7.0
djangorestframework
python-dotenv
django-cors-headers
psycopg2-binary  
djangorestframework-simplejwt

